requirements.txt
flask
google-genai
smartsheet-python-sdk
python-dotenv
psycopg2-binary
gspread
authlib
requests
flask-login
PyPDF2
python-docx
python-pptx
openpyxl
pandas
flask-wtf
cryptography
Flask-Limiter
gunicorn